+++ src/help.ts
@@ -1,6 +1,6 @@
 /**
  * This file is part of the vscode-powertools distribution.
- * Copyright (c) e.GO Digital GmbH, Aachen, Germany (https://www.e-go-digital.com/)
+ * Copyright (c) Next.e.GO Mobile SE, Aachen, Germany (https://www.e-go-mobile.com/)
  *
  * vscode-powertools is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -79,7 +79,7 @@ export class HelpViewView extends ego_webview.WebViewWithContextBase {
                             .split(' ')
                             .join('-');
                         if ('' !== pageName) {
-                            let url = `https://github.com/egodigital/vscode-powertools/wiki/${ encodeURIComponent(pageName) }.md`;
+                            let url = `https://github.com/egomobile/vscode-powertools/wiki/${ encodeURIComponent(pageName) }.md`;
 
                             let depth = 0;
                             let response: ego_helpers.HTTPRequestResult;