+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log (vscode-powertools)
 
+## 0.65.0 (June 13th, 2021; e.GO Mobile)
+
+* extension is now available at new location at [Visual Studio Code Marketplace](https://marketplace.visualstudio.com/items?itemName=ego-digital.vscode-powertools)
+* extension requires at least [Visual Studio Code 1.57](https://code.visualstudio.com/updates/v1_57) now
+* bugfixes
+* `npm update`
+
 ## 0.64.0 (February 8th, 2020; Visual Studio Code 1.53)
 
 * extension requires at least [Visual Studio Code 1.53](https://code.visualstudio.com/updates/v1_53) now